requirements.txt
# internal requirement
#-e .
#
# external requirements
Sphinx
coverage
awscli
python-dotenv>=0.5.1

# backwards compatibility
pathlib2
opencv-python==3.2.0.6
h5py>=2.6.0       #===2.6.0
matplotlib>=2.0.0 #==2.0.0
numpy>=1.13.2     #==1.112.0
scipy>=0.18.1     #==0.18.1
tqdm==4.11.2
keras==2.0.2
scikit-learn>=0.18.1 #==0.18.1
pillow>=4.0.0        #==4.0.0
scikit-image
#tensorflow>=1.6.0
